fix : CLI tests move to test_cli.py
